Prueba

+++ src/main/java/Modelos/MServicio.java
@@ -17,7 +17,7 @@ public class MServicio {
         return costo;
     }
 
-    public void setCosto(float costo) {
+    public void setCosto(double costo) {
         this.costo = costo;
     }
